+++ fastcrypto-tbls/src/tests/polynomial_tests.rs
@@ -215,7 +215,7 @@ mod scalar_tests {
         let mut rng = rand::thread_rng();
         let polynomial: Poly<S> = Poly::rand(t, &mut rng);
         let evaluations = polynomial.eval_range(n).unwrap();
-        assert_eq!(evaluations.len(), n);
+        assert_eq!(evaluations.len(), n as usize);
         for Eval { index, value } in evaluations {
             assert_eq!(value, polynomial.eval(index).value);
         }
